+++ R/badger.R
@@ -90,7 +90,7 @@ cr_deploy_badger <- function(badger_image = "gcr.io/hightowerlabs/badger:0.0.1",
 #' \code{cr_build_logs_badger} is intended to be run from the root directory of
 #' an R package that holds a README.md file containing a \code{![Cloudbuild]}
 #' badge as created by \code{cr_deploy_badger()}.  The function will scan the
-#' README.md file for the correct triggerId to pass to \link{cr_build_logs_last}
+#' README.md file for the correct triggerId to pass to \link{cr_buildtrigger_logs}
 cr_build_logs_badger <- function(dir = getwd(), projectId = cr_project_get()) {
   read_file <- file.path(dir, "README.md")
   if (!file.exists(read_file)) {
@@ -108,7 +108,7 @@ cr_build_logs_badger <- function(dir = getwd(), projectId = cr_project_get()) {
   trigger_id <- strsplit(badger_line, "&id=")[[1]][[2]]
   trigger_id <- strsplit(trigger_id, ")")[[1]]
 
-  bs <- cr_build_logs_last(trigger_id = trigger_id, projectId = projectId)
+  bs <- cr_buildtrigger_logs(trigger_id = trigger_id, projectId = projectId)
 
   invisible(bs)
 }