+++ R/cr_regions.R
@@ -2,8 +2,3 @@
 #'
 #' @format A character vector of valid Cloud Run region names
 "cr_regions"
-
-#' Cloud Run Zones
-#'
-#' @format A character vector of valid Cloud Run zone names
-"cr_zones"