+++ NEWS.md
@@ -4,7 +4,7 @@
 * Add direct Secret Manager environment args in builds via `availableSecrets` (#106)
 * Add support for `logsBucket` in `cr_build_yaml` and `cr_build_make`
 * Add `cr_build_list()` and `cr_build_list_filter()` 
-* Add `cr_build_logs()` and `cr_build_logs_last()` and `cr_build_logs_badger()`
+* Add `cr_build_logs()` and `cr_buildtrigger_logs()` and `cr_build_logs_badger()`
 * Made Cloud Build status messages prettier
 * Add messaging to `cr_build_upload_gcs()` saying where deploy folder is (#110) and clean up tar.gz folder correctly
 * Add `gcloud_args` to `cr_buildstep_run()` and `cr_run()` and `cr_deploy_run()` to allow more customisation (#113)